nav bar refactor components
