+++ booking-app/app/layout.tsx
@@ -5,7 +5,7 @@ import "@/components/src/client/styles.css";
 import { AuthProvider } from "@/components/src/client/routes/components/AuthProvider";
 import ClientProvider from "@/components/src/client/routes/components/ClientProvider";
 import CssBaseline from "@mui/material/CssBaseline";
-import NavBar from "@/components/src/client/routes/components/navBar";
+import NavBar from "@/components/src/client/routes/components/navBar/NavBar";
 import { Roboto } from "next/font/google";
 import { ThemeProvider } from "@mui/material";
 import theme from "./theme/theme";
